LDEV-2660 CFZIP action="unzip" overwrite="true" deletes existing directories (#7)

+++ source/java/src/org/lucee/extension/zip/tag/Zip.java
@@ -525,7 +525,12 @@ public final class Zip extends BodyTagImpl {
 					continue;
 				}
 				target = ZipUtil.toResource(destination, fh.getFileName());
-
+				
+				// entrypath
+				if (!entryPathMatch(path)) {
+					continue;
+				}
+				
 				if (target.exists() && overwrite) target.delete();
 
 				// filter
@@ -533,10 +538,6 @@ public final class Zip extends BodyTagImpl {
 					continue;
 				}
 
-				// entrypath
-				if (!entryPathMatch(path)) {
-					continue;
-				}
 				if (!storePath) {
 					target = destination.getRealResource(target.getName());
 				}
